Created a base SSMF model builder script with associated functions.
Created 2 optimization scripts to achieve required stability and drift values using the base SSMF model builder.

Verified base model builder works in an optimization loop & also when used in a stand-alone model.

Also modified several helper functions to account for the use of the base SSMF builder in an optimization loop or just as is for a single model.